Added unit tests for Match Facade
Through writing the Unit Tests I have discovered that I forgot to bring the properties from the Score and Round Managers. Also through testing I have run into a bunch of issues with the Singletons not properly clearing in between the tests. Going to have to do some refactoring and stop supporting the Singleton Design Pattern in this instance.