Continuing work on adding new features
Currently the user guessing is done programatically, this will all be removed once the UI is built for the user to provide input. You can actually see a lot of comments throughout the code that references areas that will be removed in a future update. The largest of the changes has been placed into main, I might look at pulling some of this code out in a future update to tighten up this class.